refactor: move test match configuration into common helper
