+++ src/scripts/__mocks__/@hover/javascript/eslint.js
@@ -1,0 +1,1 @@
+module.exports = {}
